ragSimple/requirements.txt
langchain==0.1.17
langchain_openai==0.1.5
python-dotenv==1.0.1
streamlit==1.33.0
